backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Валидация
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Авторизация / безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Telegram авторизация
hmac
hashlib

# Очереди / кэш
celery[redis]==5.4.0
redis==5.1.1

# HTTP клиент (интеграции ФНС, IRS, CBR)
httpx==0.27.2

# PDF / OCR
pypdf==5.0.1
pdf2image==1.17.0
Pillow==11.0.0

# Excel / CSV
openpyxl==3.1.5
pandas==2.2.3

# XML (3-НДФЛ)
lxml==5.3.0

# Шифрование полей
cryptography==43.0.3

# Утилиты
python-multipart==0.0.12
python-dotenv==1.0.1
structlog==24.4.0
sentry-sdk[fastapi]==2.17.0
